app/copilot-agent/requirements.txt
azure-ai-inference>=1.0.0b9
